% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pkgSHC.R
\docType{package}
\name{pkgSHC-package}
\alias{pkgSHC}
\alias{pkgSHC-package}
\title{Statistical Significance for Hierarchical Clustering}
\description{
This package implements a significance testing procedure for hierarchical
clustering based on the assumption that a cluster may be well approximated as a
single Gaussian distribution. The procedure is particularly well-suited for
the high-dimension low-sample size setting when the number of dimensions
greatly outnumbers the number of observations.
A more complete and up-to-date version of this package is actively maintained
at: http://github.com/pkimes/sigclust2 .
}
\author{
Patrick Kimes
}
\keyword{package}

